Back to unsigned int and int for {u,}intptr_t, otherwise java has no chance...

+++ sys/arch/i386/include/int_types.h
@@ -1,4 +1,4 @@
-/*	$NetBSD: int_types.h,v 1.6 2001/04/28 15:41:31 kleink Exp $	*/
+/*	$NetBSD: int_types.h,v 1.7 2003/05/29 18:06:55 christos Exp $	*/
 
 /*-
  * Copyright (c) 1990 The Regents of the University of California.
@@ -64,7 +64,7 @@ typedef	unsigned long long int __uint64_t;
 
 /* 7.18.1.4 Integer types capable of holding object pointers */
 
-typedef	long		       __intptr_t;
-typedef	unsigned long	      __uintptr_t;
+typedef	int		       __intptr_t;
+typedef	unsigned int	      __uintptr_t;
 
 #endif	/* !_I386_INT_TYPES_H_ */
